Move dead code out of sw/src/main.rs
