+++ sw/src/main.rs
@@ -26,8 +26,8 @@ use core::panic::PanicInfo;
 use gyro_rs::hal_gyro::BtGyro;
 use riscv_rt::entry;
 use utralib::generated::{
-    utra, CSR, HW_COM_BASE, HW_CRG_BASE, HW_GIT_BASE, HW_POWER_BASE, HW_SPIFLASH_MEM,
-    HW_TICKTIMER_BASE, HW_WIFI_BASE,
+    utra, CSR, HW_COM_BASE, HW_CRG_BASE, HW_GIT_BASE, HW_POWER_BASE, HW_TICKTIMER_BASE,
+    HW_WIFI_BASE,
 };
 use volatile::Volatile;
 use wfx_bindings::SL_STATUS_OK;
@@ -97,26 +97,6 @@ fn ticktimer_int_handler(_irq_no: usize) {
     set_msleep_target_ticks(50); // resetting this will also clear the alarm
 
     ticktimer_csr.wfo(utra::ticktimer::EV_PENDING_ALARM, 1);
-}
-
-#[allow(dead_code)] // used for debugging
-fn dump_rom_addr(addr: u32) {
-    let rom_ptr: *mut u32 = (addr + HW_SPIFLASH_MEM as u32) as *mut u32;
-    let rom = rom_ptr as *mut Volatile<u32>;
-    for i in 0..64 {
-        if i % 8 == 0 {
-            sprint!("\n\r0x{:06x}: ", addr + i * 4);
-        }
-        let data: u32 = unsafe { (*rom.add(i as usize)).read() };
-        sprint!(
-            "{:02x} {:02x} {:02x} {:02x} ",
-            data & 0xFF,
-            (data >> 8) & 0xff,
-            (data >> 16) & 0xff,
-            (data >> 24) & 0xff
-        );
-    }
-    sprintln!("");
 }
 
 fn ll_debug(msg: &str) {
@@ -208,31 +188,6 @@ fn main() -> ! {
     hw.charger.update_regs(&mut i2c);
     ll_debug("charger.update_regs");
 
-    /*
-    // check that the gas gauge capacity is correct; if not, reset it
-    if gg_set_design_capacity(&mut i2c, None) != 1100 {
-        gg_set_design_capacity(&mut i2c, Some(1100));
-    } */
-    // seems to work better with the default 1340mAh capacity even though that's not our actual capacity
-
-    /*  // kept around as a quick test routine for SPI flashing
-        let mut idcode: [u8; 3] = [0; 3];
-        spi_cmd(CMD_RDID, None, Some(&mut idcode));
-        sprintln!("SPI ID code: {:02x} {:02x} {:02x}", idcode[0], idcode[1], idcode[2]);
-        let test_addr = 0x8_0000;
-        dump_rom_addr(test_addr);
-        spi_erase_region(test_addr, 4096);
-
-        dump_rom_addr(test_addr);
-
-        let mut test_data: [u8; 256] = [0; 256];
-        for i in 0..256 {
-            test_data[i] = (255 - i) as u8;
-        }
-        spi_program_page(test_addr, &mut test_data);
-
-        dump_rom_addr(test_addr);
-    */
     spi_standby(); // make sure the OE's are off, no spurious power consumption
     ll_debug("spi_standby");
 
@@ -451,7 +406,8 @@ fn main() -> ! {
             } else if rx >= ComState::BL_START.verb && rx <= ComState::BL_END.verb {
                 let main_bl_level: u8 = (rx & 0x1F) as u8;
                 let sec_bl_level: u8 = ((rx >> 5) & 0x1F) as u8;
-                hw.backlight.set_brightness(&mut i2c, main_bl_level, sec_bl_level);
+                hw.backlight
+                    .set_brightness(&mut i2c, main_bl_level, sec_bl_level);
             } else if rx == ComState::LINK_READ.verb {
                 // this a "read continuation" command, in other words, return read data
                 // based on the current ComState
